requirements.txt
chromadb~=0.6.3
langchain~=0.3.20
docx2txt~=0.8
bs4~=0.0.2
beautifulsoup4~=4.13.3
uvicorn~=0.34.0
fastapi~=0.115.11
pydantic~=2.10.6
python-dotenv~=1.0.1
atlassian-python-api~=3.41.19
PyMuPDF~=1.25.3